LayoutTests/fast/page-color-sampling/color-sampling-ignores-text-expected.txt

PASS sampledTopColors[0] is "rgb(255, 100, 0)"
PASS sampledTopColors[1] is "rgb(255, 100, 0)"
PASS sampledTopColors[2] is "rgb(255, 100, 0)"
PASS sampledTopColors[3] is "rgb(255, 100, 0)"
PASS sampledTopColors[4] is "rgb(255, 100, 0)"
PASS sampledTopColors[5] is "rgb(255, 100, 0)"
PASS sampledTopColors[6] is "rgb(255, 100, 0)"
PASS sampledTopColors[7] is "rgb(255, 100, 0)"
PASS sampledTopColors[8] is "rgb(255, 100, 0)"
PASS sampledTopColors[9] is "rgb(255, 100, 0)"
PASS successfullyParsed is true

TEST COMPLETE
Here’s to the crazy ones. The misfits. The rebels. The troublemakers. The round pegs in the square holes. The ones who see things differently. They’re not fond of rules. And they have no respect for the status quo. You can quote them, disagree with them, glorify or vilify them. About the only thing you can’t do is ignore them. Because they change things. They push the human race forward. And while some may see them as the crazy ones, we see genius. Because the people who are crazy enough to think they can change the world, are the ones who do.
